Phase-II__Todo-Full-Stack-Web-Application/backend/requirements.txt
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
pytest
httpx
sqlmodel